pyproject.toml
[project]
name = "carb-scraper"
version = "0.1.0"
description = "CARB Regulations Crawler + Chatbot"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "anthropic>=0.34",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
    "beautifulsoup4>=4.12",
]
